IP Configuration in C#
This is a utility program or an application developed using C# to change the IP Configurations of the system. That is, If the system has static IP configuration, this program will change it to dynamic IP configuration and vice versa. Moreover, The application is taking the input from App.config.